+++ src/ScopelyBattles.Shared/Leaderboard/LeaderboardEntry.cs
@@ -4,5 +4,5 @@ public sealed record LeaderboardEntry
 {
     public required int Id { get; init; }
     public required string Name { get; init; }
-    public required int Score { get; init; }
+    public required long Score { get; init; }
 }
